refactor: consolidate mock data into centralized lib/mock/ with factory pattern
