+++ components/bounty-detail/bounty-detail-client.tsx
@@ -24,7 +24,7 @@ import { MilestoneFunnel } from "@/components/bounty/milestone-funnel";
 import {
   MOCK_MODEL4_MILESTONES,
   MOCK_MODEL4_CONTRIBUTORS,
-} from "@/lib/mock-model4";
+} from "@/lib/mock/model4";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { MilestoneSubmissionCard } from "./milestone-submission-card";
 import { Model4MaintainerDashboard } from "./model4-maintainer-dashboard";
