+++ lib/mock/bounties.ts
@@ -1,0 +1,281 @@
+import {
+  Bounty,
+  BountyType,
+  BountyStatus,
+  BountyOrganization,
+  BountyProject,
+} from "@/types/bounty";
+
+const defaultOrganization: BountyOrganization = {
+  id: "org-boundless",
+  name: "Boundless",
+  logo: "/logo.svg",
+  slug: "boundless",
+};
+
+const defaultProject: BountyProject = {
+  id: "proj-bounties",
+  title: "Bounties Platform",
+  description: null,
+};
+
+export function makeMockBounty(overrides: Partial<Bounty> = {}): Bounty {
+  return {
+    id: "mock-bounty-id",
+    type: "FIXED_PRICE" as BountyType,
+    title: "Mock Bounty Title",
+    organizationId: defaultOrganization.id,
+    organization: defaultOrganization,
+    projectId: defaultProject.id,
+    project: defaultProject,
+    githubIssueUrl: "https://github.com/example/repo/issues/1",
+    githubIssueNumber: 1,
+    description:
+      "## Mock Bounty\n\nThis is a mock bounty for testing purposes.",
+    rewardAmount: 100,
+    rewardCurrency: "USDC",
+    status: "OPEN" as BountyStatus,
+    createdBy: "user-admin",
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+    ...overrides,
+  };
+}
+
+export function makeMockBounties(): Bounty[] {
+  const organizations: BountyOrganization[] = [
+    defaultOrganization,
+    {
+      id: "org-soroban-kit",
+      name: "Soroban Kit",
+      logo: null,
+      slug: "soroban-kit",
+    },
+    {
+      id: "org-privacy-lab",
+      name: "Stellar Privacy Lab",
+      logo: null,
+      slug: "stellar-privacy-lab",
+    },
+  ];
+
+  const projects: BountyProject[] = [
+    defaultProject,
+    { id: "proj-platform", title: "Platform", description: null },
+    { id: "proj-sdk", title: "SDK", description: null },
+    { id: "proj-zkp", title: "ZKP", description: null },
+  ];
+
+  const bounties: Bounty[] = [
+    {
+      id: "1",
+      type: "FIXED_PRICE",
+      title: "Fix authentication redirect loop on Safari",
+      organizationId: "org-boundless",
+      organization: organizations[0],
+      projectId: "proj-bounties",
+      project: defaultProject,
+      githubIssueUrl: "https://github.com/boundlessfi/bounties/issues/1",
+      githubIssueNumber: 1,
+      description: `## Problem\nUsers on Safari are experiencing an infinite redirect loop when attempting to authenticate via GitHub OAuth.\n\n## Steps to Reproduce\n1. Open Safari browser\n2. Navigate to the login page\n3. Click "Sign in with GitHub"\n4. After GitHub authorization, observe the redirect loop\n\n## Expected Behavior\nUser should be redirected to the dashboard after successful authentication.\n\n## Technical Notes\nThe issue appears to be related to cookie handling in Safari's strict privacy mode.`,
+      rewardAmount: 500,
+      rewardCurrency: "USDC",
+      status: "OPEN",
+      createdBy: "user-admin",
+      createdAt: "2025-01-15T10:00:00Z",
+      updatedAt: "2025-01-18T14:30:00Z",
+    },
+    {
+      id: "2",
+      type: "FIXED_PRICE",
+      title: "Implement dark mode toggle in settings",
+      organizationId: "org-boundless",
+      organization: organizations[0],
+      projectId: "proj-bounties",
+      project: defaultProject,
+      githubIssueUrl: "https://github.com/boundlessfi/bounties/issues/2",
+      githubIssueNumber: 2,
+      description: `## Feature Request\nAdd a dark mode toggle to the application settings page that persists user preference.\n\n## Requirements\n- Toggle switch in settings\n- Persist preference in localStorage\n- Respect system preference as default\n- Smooth transition between themes`,
+      rewardAmount: 300,
+      rewardCurrency: "USDC",
+      status: "IN_PROGRESS",
+      createdBy: "user-admin",
+      createdAt: "2025-01-10T08:00:00Z",
+      updatedAt: "2025-01-17T11:00:00Z",
+    },
+    {
+      id: "3",
+      type: "FIXED_PRICE",
+      title: "Write API documentation for bounty endpoints",
+      organizationId: "org-boundless",
+      organization: organizations[0],
+      projectId: "proj-bounties",
+      project: defaultProject,
+      githubIssueUrl: "https://github.com/boundlessfi/bounties/issues/3",
+      githubIssueNumber: 3,
+      description: `## Documentation Needed\nCreate comprehensive API documentation for all bounty-related endpoints.\n\n## Endpoints to Document\n- GET /api/bounties\n- GET /api/bounties/:id\n- POST /api/bounties/:id/claim\n- POST /api/bounties/:id/submit`,
+      rewardAmount: 200,
+      rewardCurrency: "USDC",
+      status: "COMPLETED",
+      createdBy: "user-admin",
+      createdAt: "2025-01-05T12:00:00Z",
+      updatedAt: "2025-01-12T16:00:00Z",
+    },
+    {
+      id: "4",
+      type: "COMPETITION",
+      title: "Add wallet connection support",
+      organizationId: "org-boundless",
+      organization: organizations[0],
+      projectId: "proj-platform",
+      project: projects[1],
+      githubIssueUrl: "https://github.com/boundlessfi/platform/issues/45",
+      githubIssueNumber: 45,
+      description: `## Feature Request\nImplement wallet connection functionality using Freighter and other popular Stellar wallets.\n\n## Requirements\n- Support multiple wallet providers\n- Handle connection/disconnection flows\n- Display connected wallet address\n- Persist wallet connection state`,
+      rewardAmount: 800,
+      rewardCurrency: "XLM",
+      status: "OPEN",
+      createdBy: "user-admin",
+      createdAt: "2025-01-20T09:00:00Z",
+      updatedAt: "2025-01-23T11:30:00Z",
+    },
+    {
+      id: "5",
+      type: "FIXED_PRICE",
+      title: "Fix pagination on bounties list",
+      organizationId: "org-boundless",
+      organization: organizations[0],
+      projectId: "proj-platform",
+      project: projects[1],
+      githubIssueUrl: "https://github.com/boundlessfi/platform/issues/52",
+      githubIssueNumber: 52,
+      description: `## Bug Description\nPagination controls are not working correctly when filtering bounties.\n\n## Steps to Reproduce\n1. Apply filters to bounties list\n2. Navigate to page 2\n3. Change filter\n4. Pagination state is not reset`,
+      rewardAmount: 150,
+      rewardCurrency: "USDC",
+      status: "OPEN",
+      createdBy: "user-admin",
+      createdAt: "2025-01-22T14:00:00Z",
+      updatedAt: "2025-01-23T10:15:00Z",
+    },
+    {
+      id: "6",
+      type: "MILESTONE_BASED",
+      title: "Refactor API client to use axios",
+      organizationId: "org-boundless",
+      organization: organizations[0],
+      projectId: "proj-platform",
+      project: projects[1],
+      githubIssueUrl: "https://github.com/boundlessfi/platform/issues/38",
+      githubIssueNumber: 38,
+      description: `## Refactoring Task\nReplace current fetch-based API client with axios for better error handling and interceptor support.`,
+      rewardAmount: 400,
+      rewardCurrency: "USDC",
+      status: "IN_PROGRESS",
+      createdBy: "user-admin",
+      createdAt: "2025-01-18T08:00:00Z",
+      updatedAt: "2025-01-21T16:45:00Z",
+    },
+    {
+      id: "7",
+      type: "MILESTONE_BASED",
+      title: "Implement Soroban smart contract integration",
+      organizationId: "org-soroban-kit",
+      organization: organizations[1],
+      projectId: "proj-sdk",
+      project: projects[2],
+      githubIssueUrl: "https://github.com/soroban-kit/sdk/issues/12",
+      githubIssueNumber: 12,
+      description: `## Feature Request\nAdd helper functions for common Soroban smart contract patterns.\n\n## Scope\n- Token contract interactions\n- Authorization patterns\n- Error handling utilities\n- Testing helpers`,
+      rewardAmount: 1200,
+      rewardCurrency: "XLM",
+      status: "OPEN",
+      createdBy: "user-admin",
+      createdAt: "2025-01-19T10:30:00Z",
+      updatedAt: "2025-01-24T09:20:00Z",
+    },
+    {
+      id: "8",
+      type: "FIXED_PRICE",
+      title: "Fix contract deployment script errors",
+      organizationId: "org-soroban-kit",
+      organization: organizations[1],
+      projectId: "proj-sdk",
+      project: projects[2],
+      githubIssueUrl: "https://github.com/soroban-kit/sdk/issues/18",
+      githubIssueNumber: 18,
+      description: `## Bug Report\nDeployment scripts fail on testnet with network timeout errors.\n\n## Environment\n- Network: Stellar Testnet\n- Node version: 18.x\n- Soroban CLI: 20.0.0`,
+      rewardAmount: 300,
+      rewardCurrency: "USDC",
+      status: "OPEN",
+      createdBy: "user-admin",
+      createdAt: "2025-01-21T11:00:00Z",
+      updatedAt: "2025-01-23T15:30:00Z",
+    },
+    {
+      id: "9",
+      type: "FIXED_PRICE",
+      title: "Create getting started guide",
+      organizationId: "org-soroban-kit",
+      organization: organizations[1],
+      projectId: "proj-sdk",
+      project: projects[2],
+      githubIssueUrl: "https://github.com/soroban-kit/sdk/issues/5",
+      githubIssueNumber: 5,
+      description: `## Documentation Request\nWrite comprehensive getting started guide for new developers.\n\n## Sections Needed\n- Installation\n- Quick start example\n- Core concepts\n- Common patterns\n- Troubleshooting`,
+      rewardAmount: 250,
+      rewardCurrency: "USDC",
+      status: "OPEN",
+      createdBy: "user-admin",
+      createdAt: "2025-01-15T13:00:00Z",
+      updatedAt: "2025-01-22T10:00:00Z",
+    },
+    {
+      id: "10",
+      type: "COMPETITION",
+      title: "Add zero-knowledge proof primitives",
+      organizationId: "org-privacy-lab",
+      organization: organizations[2],
+      projectId: "proj-zkp",
+      project: projects[3],
+      githubIssueUrl: "https://github.com/stellar-privacy/zkp/issues/3",
+      githubIssueNumber: 3,
+      description: `## Research Feature\nImplement basic zero-knowledge proof primitives for private transactions on Stellar.\n\n## Requirements\n- Research ZK-SNARK implementations\n- Prototype basic proof generation\n- Integration with Stellar transactions\n- Performance benchmarks`,
+      rewardAmount: 2000,
+      rewardCurrency: "XLM",
+      status: "OPEN",
+      createdBy: "user-admin",
+      createdAt: "2025-01-10T09:00:00Z",
+      updatedAt: "2025-01-24T14:20:00Z",
+    },
+    {
+      id: "11",
+      type: "FIXED_PRICE",
+      title: "Fix privacy protocol memory leaks",
+      organizationId: "org-privacy-lab",
+      organization: organizations[2],
+      projectId: "proj-sdk",
+      project: projects[2],
+      githubIssueUrl: "https://github.com/stellar-privacy/sdk/issues/7",
+      githubIssueNumber: 7,
+      description: `## Bug Description\nMemory leaks detected in privacy protocol implementation during extended operations.`,
+      rewardAmount: 600,
+      rewardCurrency: "USDC",
+      status: "OPEN",
+      createdBy: "user-admin",
+      createdAt: "2025-01-19T16:00:00Z",
+      updatedAt: "2025-01-23T12:45:00Z",
+    },
+  ];
+
+  return bounties;
+}
+
+export const mockBounties = makeMockBounties();
+
+export function getBountyById(id: string): Bounty | undefined {
+  return mockBounties.find((bounty) => bounty.id === id);
+}
+
+export function getAllBounties(): Bounty[] {
+  return mockBounties;
+}
