+++ lib/mock/discover.ts
@@ -1,0 +1,508 @@
+import type { Project } from "@/lib/types";
+import type { Bounty } from "@/types/bounty";
+
+export function makeOldProject(overrides: Partial<Project> = {}): Project {
+  return {
+    id: "proj-mock",
+    title: "Mock Project",
+    description: "Mock project description",
+    tags: [],
+    status: "active",
+    createdAt: new Date().toISOString(),
+    updatedAt: new Date().toISOString(),
+    creator: "mock-creator",
+    category: "Mock",
+    milestones: 0,
+    completedMilestones: 0,
+    ...overrides,
+  };
+}
+
+export function makeOldProjects(): Project[] {
+  return [
+    {
+      id: "proj-1",
+      title: "Stellar DeFi Dashboard",
+      description:
+        "A comprehensive dashboard for tracking DeFi protocols on Stellar network. Features real-time analytics, portfolio tracking, and yield optimization.",
+      tags: ["DeFi", "Frontend", "Analytics", "Stellar"],
+      status: "active",
+      createdAt: "2024-01-15T00:00:00Z",
+      updatedAt: "2024-01-20T00:00:00Z",
+      creator: "stellar_dev",
+      category: "DeFi",
+      milestones: 5,
+      completedMilestones: 3,
+    },
+    {
+      id: "proj-2",
+      title: "NFT Marketplace on Stellar",
+      description:
+        "Decentralized NFT marketplace built on Stellar. Supports minting, trading, and royalty management with low transaction fees.",
+      tags: ["NFT", "Smart Contracts", "Full Stack", "Stellar"],
+      status: "active",
+      createdAt: "2024-01-10T00:00:00Z",
+      updatedAt: "2024-01-22T00:00:00Z",
+      creator: "nft_builder",
+      category: "NFT",
+      milestones: 8,
+      completedMilestones: 5,
+    },
+    {
+      id: "proj-3",
+      title: "Cross-Chain Bridge Protocol",
+      description:
+        "Secure bridge protocol enabling asset transfers between Stellar and other major blockchains.",
+      tags: ["DeFi", "Smart Contracts", "Security", "Infrastructure"],
+      status: "active",
+      createdAt: "2024-01-05T00:00:00Z",
+      updatedAt: "2024-01-18T00:00:00Z",
+      creator: "bridge_team",
+      category: "Infrastructure",
+      milestones: 6,
+      completedMilestones: 2,
+    },
+    {
+      id: "proj-4",
+      title: "Stellar Mobile Wallet",
+      description:
+        "User-friendly mobile wallet for Stellar assets with built-in DEX integration and staking features.",
+      tags: ["Mobile", "Frontend", "Web3", "Stellar"],
+      status: "active",
+      createdAt: "2023-12-20T00:00:00Z",
+      updatedAt: "2024-01-21T00:00:00Z",
+      creator: "mobile_dev",
+      category: "Wallet",
+      milestones: 10,
+      completedMilestones: 8,
+    },
+    {
+      id: "proj-5",
+      title: "DAO Governance Platform",
+      description:
+        "Decentralized governance platform for DAOs on Stellar with voting mechanisms and proposal management.",
+      tags: ["Smart Contracts", "Frontend", "Backend", "Web3"],
+      status: "completed",
+      createdAt: "2023-11-01T00:00:00Z",
+      updatedAt: "2023-12-15T00:00:00Z",
+      creator: "dao_builders",
+      category: "Governance",
+      milestones: 4,
+      completedMilestones: 4,
+    },
+    {
+      id: "proj-6",
+      title: "Stellar Analytics Engine",
+      description:
+        "Advanced analytics engine for Stellar blockchain data with customizable dashboards and alerts.",
+      tags: ["Analytics", "Backend", "Infrastructure", "Stellar"],
+      status: "active",
+      createdAt: "2024-01-12T00:00:00Z",
+      updatedAt: "2024-01-19T00:00:00Z",
+      creator: "analytics_pro",
+      category: "Analytics",
+      milestones: 7,
+      completedMilestones: 4,
+    },
+    {
+      id: "proj-7",
+      title: "Smart Contract Testing Suite",
+      description:
+        "Comprehensive testing framework for Stellar smart contracts with automated security audits.",
+      tags: ["Testing", "Security", "Smart Contracts", "Infrastructure"],
+      status: "paused",
+      createdAt: "2023-12-01T00:00:00Z",
+      updatedAt: "2024-01-10T00:00:00Z",
+      creator: "test_master",
+      category: "Development Tools",
+      milestones: 5,
+      completedMilestones: 2,
+    },
+    {
+      id: "proj-8",
+      title: "Decentralized Identity System",
+      description:
+        "Self-sovereign identity solution on Stellar for secure credential management and verification.",
+      tags: ["Security", "Smart Contracts", "Backend", "Web3"],
+      status: "active",
+      createdAt: "2024-01-08T00:00:00Z",
+      updatedAt: "2024-01-22T00:00:00Z",
+      creator: "identity_dev",
+      category: "Identity",
+      milestones: 6,
+      completedMilestones: 3,
+    },
+    {
+      id: "proj-9",
+      title: "Stellar Documentation Hub",
+      description:
+        "Comprehensive documentation platform with interactive tutorials and code examples for Stellar developers.",
+      tags: ["Documentation", "Frontend", "Design"],
+      status: "completed",
+      createdAt: "2023-10-15T00:00:00Z",
+      updatedAt: "2023-12-01T00:00:00Z",
+      creator: "docs_team",
+      category: "Education",
+      milestones: 3,
+      completedMilestones: 3,
+    },
+    {
+      id: "proj-10",
+      title: "Yield Aggregator Protocol",
+      description:
+        "Automated yield optimization protocol that finds the best returns across Stellar DeFi platforms.",
+      tags: ["DeFi", "Smart Contracts", "Backend", "Analytics"],
+      status: "active",
+      createdAt: "2024-01-14T00:00:00Z",
+      updatedAt: "2024-01-21T00:00:00Z",
+      creator: "yield_hunter",
+      category: "DeFi",
+      milestones: 8,
+      completedMilestones: 4,
+    },
+  ];
+}
+
+export function makeOldBounties(projects: Project[]): Bounty[] {
+  return [
+    {
+      id: "bounty-1",
+      title: "Implement Multi-Signature Wallet Feature",
+      description:
+        "Add multi-signature functionality to existing Stellar wallet with customizable approval thresholds.",
+      type: "FIXED_PRICE",
+      rewardAmount: 5000,
+      rewardCurrency: "USDC",
+      status: "OPEN",
+      organizationId: "org-wallet",
+      organization: {
+        id: "org-wallet",
+        name: "Wallet Project",
+        logo: "/logos/org-wallet.png",
+        slug: "org-wallet",
+      },
+      projectId: projects[0].id,
+      project: {
+        id: projects[0].id,
+        title: projects[0].title,
+        description: null,
+      },
+      githubIssueUrl: "https://github.com/wallet-project/issues/1",
+      githubIssueNumber: null,
+      createdBy: "wallet_project",
+      createdAt: "2024-01-20T00:00:00Z",
+      updatedAt: "2024-01-20T00:00:00Z",
+    },
+    {
+      id: "bounty-2",
+      title: "Design Landing Page for DeFi Protocol",
+      description:
+        "Create modern, responsive landing page design with dark mode support and animated elements.",
+      type: "COMPETITION",
+      rewardAmount: 2000,
+      rewardCurrency: "USDC",
+      status: "OPEN",
+      organizationId: "org-defi",
+      organization: {
+        id: "org-defi",
+        name: "DeFi Startup",
+        logo: "/logos/org-defi.png",
+        slug: "org-defi",
+      },
+      projectId: projects[1].id,
+      project: {
+        id: projects[1].id,
+        title: projects[1].title,
+        description: null,
+      },
+      githubIssueUrl: "https://github.com/defi-startup/issues/1",
+      githubIssueNumber: null,
+      createdBy: "defi_startup",
+      createdAt: "2024-01-19T00:00:00Z",
+      updatedAt: "2024-01-21T00:00:00Z",
+    },
+    {
+      id: "bounty-3",
+      title: "Fix Security Vulnerability in Smart Contract",
+      description:
+        "Identify and fix critical security vulnerability in liquidity pool smart contract.",
+      type: "FIXED_PRICE",
+      rewardAmount: 8000,
+      rewardCurrency: "USDC",
+      status: "IN_PROGRESS",
+      organizationId: "org-security",
+      organization: {
+        id: "org-security",
+        name: "Security Team",
+        logo: "/logos/org-security.png",
+        slug: "org-security",
+      },
+      projectId: projects[2].id,
+      project: {
+        id: projects[2].id,
+        title: projects[2].title,
+        description: null,
+      },
+      githubIssueUrl: "https://github.com/security-team/issues/1",
+      githubIssueNumber: null,
+      createdBy: "security_team",
+      createdAt: "2024-01-18T00:00:00Z",
+      updatedAt: "2024-01-22T00:00:00Z",
+    },
+    {
+      id: "bounty-4",
+      title: "Build Mobile App UI Components",
+      description:
+        "Create reusable React Native components for Stellar wallet mobile app.",
+      type: "MILESTONE_BASED",
+      rewardAmount: 3000,
+      rewardCurrency: "USDC",
+      status: "OPEN",
+      organizationId: "org-mobile",
+      organization: {
+        id: "org-mobile",
+        name: "Mobile Team",
+        logo: "/logos/org-mobile.png",
+        slug: "org-mobile",
+      },
+      projectId: projects[3].id,
+      project: {
+        id: projects[3].id,
+        title: projects[3].title,
+        description: null,
+      },
+      githubIssueUrl: "https://github.com/mobile-team/issues/1",
+      githubIssueNumber: null,
+      createdBy: "mobile_team",
+      createdAt: "2024-01-17T00:00:00Z",
+      updatedAt: "2024-01-17T00:00:00Z",
+    },
+    {
+      id: "bounty-5",
+      title: "Write Integration Tests for DEX",
+      description:
+        "Develop comprehensive integration test suite for decentralized exchange smart contracts.",
+      type: "FIXED_PRICE",
+      rewardAmount: 2500,
+      rewardCurrency: "USDC",
+      status: "OPEN",
+      organizationId: "org-dex",
+      organization: {
+        id: "org-dex",
+        name: "DEX Protocol",
+        logo: "/logos/org-dex.png",
+        slug: "org-dex",
+      },
+      projectId: projects[4].id,
+      project: {
+        id: projects[4].id,
+        title: projects[4].title,
+        description: null,
+      },
+      githubIssueUrl: "https://github.com/dex-protocol/issues/1",
+      githubIssueNumber: null,
+      createdBy: "dex_protocol",
+      createdAt: "2024-01-16T00:00:00Z",
+      updatedAt: "2024-01-20T00:00:00Z",
+    },
+    {
+      id: "bounty-6",
+      title: "Optimize Gas Fees for NFT Minting",
+      description:
+        "Reduce transaction costs for NFT minting operations by optimizing smart contract code.",
+      type: "FIXED_PRICE",
+      rewardAmount: 4000,
+      rewardCurrency: "USDC",
+      status: "COMPLETED",
+      organizationId: "org-nft",
+      organization: {
+        id: "org-nft",
+        name: "NFT Marketplace",
+        logo: "/logos/org-nft.png",
+        slug: "org-nft",
+      },
+      projectId: projects[5].id,
+      project: {
+        id: projects[5].id,
+        title: projects[5].title,
+        description: null,
+      },
+      githubIssueUrl: "https://github.com/nft-marketplace/issues/1",
+      githubIssueNumber: null,
+      createdBy: "nft_marketplace",
+      createdAt: "2024-01-10T00:00:00Z",
+      updatedAt: "2024-01-15T00:00:00Z",
+    },
+    {
+      id: "bounty-7",
+      title: "Create Tutorial Videos for Beginners",
+      description:
+        "Produce 5 tutorial videos explaining Stellar development basics for newcomers.",
+      type: "MILESTONE_BASED",
+      rewardAmount: 1500,
+      rewardCurrency: "USDC",
+      status: "OPEN",
+      organizationId: "org-edu",
+      organization: {
+        id: "org-edu",
+        name: "Education DAO",
+        logo: "/logos/org-edu.png",
+        slug: "org-edu",
+      },
+      projectId: projects[6].id,
+      project: {
+        id: projects[6].id,
+        title: projects[6].title,
+        description: null,
+      },
+      githubIssueUrl: "https://github.com/education-dao/issues/1",
+      githubIssueNumber: null,
+      createdBy: "education_dao",
+      createdAt: "2024-01-15T00:00:00Z",
+      updatedAt: "2024-01-18T00:00:00Z",
+    },
+    {
+      id: "bounty-8",
+      title: "Implement Real-Time Price Oracle",
+      description:
+        "Build reliable price oracle service for DeFi protocols with multiple data sources.",
+      type: "FIXED_PRICE",
+      rewardAmount: 6000,
+      rewardCurrency: "USDC",
+      status: "OPEN",
+      organizationId: "org-oracle",
+      organization: {
+        id: "org-oracle",
+        name: "Oracle Network",
+        logo: "/logos/org-oracle.png",
+        slug: "org-oracle",
+      },
+      projectId: projects[7].id,
+      project: {
+        id: projects[7].id,
+        title: projects[7].title,
+        description: null,
+      },
+      githubIssueUrl: "https://github.com/oracle-network/issues/1",
+      githubIssueNumber: null,
+      createdBy: "oracle_network",
+      createdAt: "2024-01-14T00:00:00Z",
+      updatedAt: "2024-01-21T00:00:00Z",
+    },
+    {
+      id: "bounty-9",
+      title: "Add Dark Mode to Dashboard",
+      description:
+        "Implement dark mode theme with smooth transitions for analytics dashboard.",
+      type: "FIXED_PRICE",
+      rewardAmount: 1000,
+      rewardCurrency: "USDC",
+      status: "IN_PROGRESS",
+      organizationId: "org-analytics",
+      organization: {
+        id: "org-analytics",
+        name: "Analytics Platform",
+        logo: "/logos/org-analytics.png",
+        slug: "org-analytics",
+      },
+      projectId: projects[8].id,
+      project: {
+        id: projects[8].id,
+        title: projects[8].title,
+        description: null,
+      },
+      githubIssueUrl: "https://github.com/analytics-platform/issues/1",
+      githubIssueNumber: null,
+      createdBy: "analytics_platform",
+      createdAt: "2024-01-13T00:00:00Z",
+      updatedAt: "2024-01-19T00:00:00Z",
+    },
+    {
+      id: "bounty-10",
+      title: "Audit Staking Contract",
+      description:
+        "Perform comprehensive security audit of staking smart contract with detailed report.",
+      type: "FIXED_PRICE",
+      rewardAmount: 7000,
+      rewardCurrency: "USDC",
+      status: "OPEN",
+      organizationId: "org-staking",
+      organization: {
+        id: "org-staking",
+        name: "Staking Protocol",
+        logo: "/logos/org-staking.png",
+        slug: "org-staking",
+      },
+      projectId: projects[9].id,
+      project: {
+        id: projects[9].id,
+        title: projects[9].title,
+        description: null,
+      },
+      githubIssueUrl: "https://github.com/staking-protocol/issues/1",
+      githubIssueNumber: null,
+      createdBy: "staking_protocol",
+      createdAt: "2024-01-12T00:00:00Z",
+      updatedAt: "2024-01-22T00:00:00Z",
+    },
+    {
+      id: "bounty-11",
+      title: "Build API Documentation Site",
+      description:
+        "Create interactive API documentation website with code examples and playground.",
+      type: "COMPETITION",
+      rewardAmount: 2800,
+      rewardCurrency: "USDC",
+      status: "OPEN",
+      organizationId: "org-api",
+      organization: {
+        id: "org-api",
+        name: "API Team",
+        logo: "/logos/org-api.png",
+        slug: "org-api",
+      },
+      projectId: projects[0].id,
+      project: {
+        id: projects[0].id,
+        title: projects[0].title,
+        description: null,
+      },
+      githubIssueUrl: "https://github.com/api-team/issues/1",
+      githubIssueNumber: null,
+      createdBy: "api_team",
+      createdAt: "2024-01-11T00:00:00Z",
+      updatedAt: "2024-01-16T00:00:00Z",
+    },
+    {
+      id: "bounty-12",
+      title: "Integrate Wallet Connect",
+      description:
+        "Add Wallet Connect support to DApp for seamless mobile wallet integration.",
+      type: "FIXED_PRICE",
+      rewardAmount: 3500,
+      rewardCurrency: "USDC",
+      status: "OPEN",
+      organizationId: "org-dapp",
+      organization: {
+        id: "org-dapp",
+        name: "DApp Builders",
+        logo: "/logos/org-dapp.png",
+        slug: "org-dapp",
+      },
+      projectId: projects[1].id,
+      project: {
+        id: projects[1].id,
+        title: projects[1].title,
+        description: null,
+      },
+      githubIssueUrl: "https://github.com/dapp-builders/issues/1",
+      githubIssueNumber: null,
+      createdBy: "dapp_builders",
+      createdAt: "2024-01-09T00:00:00Z",
+      updatedAt: "2024-01-20T00:00:00Z",
+    },
+  ];
+}
+
+export const oldMockProjects = makeOldProjects();
+export const oldMockBounties = makeOldBounties(makeOldProjects());
