+++ lib/mock/index.ts
@@ -1,0 +1,9 @@
+export * from "./bounties";
+export * from "./leaderboard";
+export * from "./wallet";
+export * from "./projects";
+export * from "./model4";
+export {
+  oldMockProjects as mockProjects,
+  oldMockBounties as mockBounties,
+} from "./discover";
