+++ lib/store.ts
@@ -1,7 +1,12 @@
 import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
 import { Bounty } from "@/types/bounty";
-import { Application, Submission, MilestoneParticipation, CompetitionParticipation } from "@/types/participation";
-import { mockBounties } from "./mock-bounty";
+import {
+  Application,
+  Submission,
+  MilestoneParticipation,
+  CompetitionParticipation,
+} from "@/types/participation";
+import { mockBounties } from "./mock/bounties";
 
 /**
  * @deprecated The previous globalThis-based local store is deprecated.
@@ -14,11 +19,16 @@ export const localStoreKeys = {
   applications: ["localStore", "applications"] as const,
   submissions: ["localStore", "submissions"] as const,
   milestoneParticipations: ["localStore", "milestoneParticipations"] as const,
-  competitionParticipations: ["localStore", "competitionParticipations"] as const,
+  competitionParticipations: [
+    "localStore",
+    "competitionParticipations",
+  ] as const,
 };
 
 function notFoundError(entity: string, id: string) {
-  return new Error(`${entity} with id "${id}" was not found in the local query cache.`);
+  return new Error(
+    `${entity} with id "${id}" was not found in the local query cache.`,
+  );
 }
 
 // --- Bounties ---
@@ -34,20 +44,29 @@ export function useLocalBounties() {
 export function useUpdateLocalBounty() {
   const queryClient = useQueryClient();
   return useMutation({
-    mutationFn: async ({ id, updates }: { id: string; updates: Partial<Bounty> }) => {
+    mutationFn: async ({
+      id,
+      updates,
+    }: {
+      id: string;
+      updates: Partial<Bounty>;
+    }) => {
       let updatedBounty: Bounty | null = null;
 
-      queryClient.setQueryData<Bounty[]>(localStoreKeys.bounties, (current = [...mockBounties]) => {
-        const index = current.findIndex((bounty) => bounty.id === id);
-        if (index === -1) {
-          return current;
-        }
+      queryClient.setQueryData<Bounty[]>(
+        localStoreKeys.bounties,
+        (current = [...mockBounties]) => {
+          const index = current.findIndex((bounty) => bounty.id === id);
+          if (index === -1) {
+            return current;
+          }
 
-        const next = [...current];
-        updatedBounty = { ...next[index], ...updates };
-        next[index] = updatedBounty;
-        return next;
-      });
+          const next = [...current];
+          updatedBounty = { ...next[index], ...updates };
+          next[index] = updatedBounty;
+          return next;
+        },
+      );
 
       if (!updatedBounty) {
         throw notFoundError("Bounty", id);
@@ -65,7 +84,8 @@ export function useLocalApplications(bountyId?: string) {
     queryFn: () => [] as Application[],
     initialData: [],
     staleTime: Infinity,
-    select: (apps) => (bountyId ? apps.filter((app) => app.bountyId === bountyId) : apps),
+    select: (apps) =>
+      bountyId ? apps.filter((app) => app.bountyId === bountyId) : apps,
   });
 }
 
@@ -73,10 +93,10 @@ export function useAddLocalApplication() {
   const queryClient = useQueryClient();
   return useMutation({
     mutationFn: async (application: Application) => {
-      queryClient.setQueryData<Application[]>(localStoreKeys.applications, (current = []) => [
-        ...current,
-        application,
-      ]);
+      queryClient.setQueryData<Application[]>(
+        localStoreKeys.applications,
+        (current = []) => [...current, application],
+      );
 
       return application;
     },
@@ -86,20 +106,31 @@ export function useAddLocalApplication() {
 export function useUpdateLocalApplication() {
   const queryClient = useQueryClient();
   return useMutation({
-    mutationFn: async ({ id, updates }: { id: string; updates: Partial<Application> }) => {
+    mutationFn: async ({
+      id,
+      updates,
+    }: {
+      id: string;
+      updates: Partial<Application>;
+    }) => {
       let updatedApplication: Application | null = null;
 
-      queryClient.setQueryData<Application[]>(localStoreKeys.applications, (current = []) => {
-        const index = current.findIndex((application) => application.id === id);
-        if (index === -1) {
-          return current;
-        }
+      queryClient.setQueryData<Application[]>(
+        localStoreKeys.applications,
+        (current = []) => {
+          const index = current.findIndex(
+            (application) => application.id === id,
+          );
+          if (index === -1) {
+            return current;
+          }
 
-        const next = [...current];
-        updatedApplication = { ...next[index], ...updates };
-        next[index] = updatedApplication;
-        return next;
-      });
+          const next = [...current];
+          updatedApplication = { ...next[index], ...updates };
+          next[index] = updatedApplication;
+          return next;
+        },
+      );
 
       if (!updatedApplication) {
         throw notFoundError("Application", id);
@@ -118,7 +149,9 @@ export function useLocalSubmissions(bountyId?: string) {
     initialData: [],
     staleTime: Infinity,
     select: (submissions) =>
-      bountyId ? submissions.filter((submission) => submission.bountyId === bountyId) : submissions,
+      bountyId
+        ? submissions.filter((submission) => submission.bountyId === bountyId)
+        : submissions,
   });
 }
 
@@ -126,10 +159,10 @@ export function useAddLocalSubmission() {
   const queryClient = useQueryClient();
   return useMutation({
     mutationFn: async (submission: Submission) => {
-      queryClient.setQueryData<Submission[]>(localStoreKeys.submissions, (current = []) => [
-        ...current,
-        submission,
-      ]);
+      queryClient.setQueryData<Submission[]>(
+        localStoreKeys.submissions,
+        (current = []) => [...current, submission],
+      );
 
       return submission;
     },
@@ -139,20 +172,29 @@ export function useAddLocalSubmission() {
 export function useUpdateLocalSubmission() {
   const queryClient = useQueryClient();
   return useMutation({
-    mutationFn: async ({ id, updates }: { id: string; updates: Partial<Submission> }) => {
+    mutationFn: async ({
+      id,
+      updates,
+    }: {
+      id: string;
+      updates: Partial<Submission>;
+    }) => {
       let updatedSubmission: Submission | null = null;
 
-      queryClient.setQueryData<Submission[]>(localStoreKeys.submissions, (current = []) => {
-        const index = current.findIndex((submission) => submission.id === id);
-        if (index === -1) {
-          return current;
-        }
+      queryClient.setQueryData<Submission[]>(
+        localStoreKeys.submissions,
+        (current = []) => {
+          const index = current.findIndex((submission) => submission.id === id);
+          if (index === -1) {
+            return current;
+          }
 
-        const next = [...current];
-        updatedSubmission = { ...next[index], ...updates };
-        next[index] = updatedSubmission;
-        return next;
-      });
+          const next = [...current];
+          updatedSubmission = { ...next[index], ...updates };
+          next[index] = updatedSubmission;
+          return next;
+        },
+      );
 
       if (!updatedSubmission) {
         throw notFoundError("Submission", id);
@@ -172,7 +214,9 @@ export function useLocalMilestoneParticipations(bountyId?: string) {
     staleTime: Infinity,
     select: (participations) =>
       bountyId
-        ? participations.filter((participation) => participation.bountyId === bountyId)
+        ? participations.filter(
+            (participation) => participation.bountyId === bountyId,
+          )
         : participations,
   });
 }
@@ -206,7 +250,9 @@ export function useUpdateLocalMilestoneParticipation() {
       queryClient.setQueryData<MilestoneParticipation[]>(
         localStoreKeys.milestoneParticipations,
         (current = []) => {
-          const index = current.findIndex((participation) => participation.id === id);
+          const index = current.findIndex(
+            (participation) => participation.id === id,
+          );
           if (index === -1) {
             return current;
           }
@@ -236,7 +282,9 @@ export function useLocalCompetitionParticipations(bountyId?: string) {
     staleTime: Infinity,
     select: (participations) =>
       bountyId
-        ? participations.filter((participation) => participation.bountyId === bountyId)
+        ? participations.filter(
+            (participation) => participation.bountyId === bountyId,
+          )
         : participations,
   });
 }
